feat: responsive redesign of Problems
Split 703-line file into Problems (226), ProblemFormModal (374), and
DeleteConfirmModal (61). Remove palette object and all inline styles.
Use Button/Input/Label UI components. Mobile-first responsive modals.
